+++ Sources/Runner/TestRunner.swift
@@ -1,17 +1,21 @@
 import Foundation
+import DeveloperDirLocator
 import Models
 import ProcessController
+import TemporaryStuff
 
 public protocol TestRunner {
     func run(
         buildArtifacts: BuildArtifacts,
+        developerDirLocator: DeveloperDirLocator,
         entriesToRun: [TestEntry],
         maximumAllowedSilenceDuration: TimeInterval,
         simulatorSettings: SimulatorSettings,
         singleTestMaximumDuration: TimeInterval,
         testContext: TestContext,
         testRunnerStream: TestRunnerStream,
-        testType: TestType
-    ) -> StandardStreamsCaptureConfig
+        testType: TestType,
+        temporaryFolder: TemporaryFolder
+    ) throws -> StandardStreamsCaptureConfig
 }